requirements.txt
praw
yt-dlp
anthropic
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
